\name{fstart}
\alias{fstart}
\alias{maxx}
\alias{minx}
\title{SAC array values}
\description{
  Calculates the starting time, the max, and the min values of \code{x}
  for a vector of SAC files.
}
\usage{
fstart(X)
maxx(X)
minx(X)
}
\arguments{
  \item{X}{A vector of SAC files.}
}
\value{the max, min, or starting time.
}
\author{Eric M. Thompson <eric.thompson@tufts.edu>}
\seealso{\code{\link{psac}} }
\examples{
# Used by psac()
}
\keyword{internal}
