\name{getunits}
\alias{getunits}
\title{SAC time series units}
\description{
  Looks up the units for when they are specified in the header as
  integer values.
}
\usage{
getunits(s)
}
\arguments{
  \item{s}{A vector of SAC files.}
}
\value{an array of strings that describe the units of the SAC files.
}
\author{Eric M. Thompson <eric.thompson@tufts.edu>}
\examples{

}
\keyword{}
