\name{rtrend}
\alias{rtrend}
\title{Remove linear trend}
\description{
  Removes the linear trend from SAC data.
}
\usage{
rtrend(s)
}
\arguments{
  \item{s}{a vector of SAC data.}
}
\value{a vector of same length as input.
}
\author{Eric M. Thompson <eric.thompson@tufts.edu>}
\examples{

}
\keyword{}
