\name{fstart}
\alias{fstart}
\alias{maxx}
\alias{minx}
\title{SAC ARRAY VALUES}
\description{
  Calculates the starting time, the max, and the min values of \code{X}
  for a vector of SAC files.
}
\usage{
fstart(X)
maxx(X)
minx(X)
}
\arguments{
  \item{X}{an object of class \code{rsac}.}
}
\value{the max, min, or starting time.
}
\author{Eric M. Thompson <eric.thompson@tufts.edu>}
\examples{
# Used by "plot.rsac"
}
\keyword{internal}
