\name{rtrend}
\alias{rtrend}
\alias{rmean}
\title{Remove linear trend or mean}
\description{
  Removes the linear trend or mean from SAC data.
}
\usage{
rtrend(x)
rmean(x)
}
\arguments{
  \item{x}{an object of class \code{"rsac"}.}
}
\value{same as input.
}
\author{Eric M. Thompson <eric.thompson@tufts.edu>}
\examples{

}
\keyword{}
