\name{rtrend}
\alias{rtrend}
\alias{rmean}
\title{Remove linear trend or mean}
\description{
  Removes the linear trend or mean from SAC data.
}
\usage{
rtrend(s)
rmean(s)
}
\arguments{
  \item{s}{a vector of SAC data.}
}
\value{a vector of same length as input.
}
\author{Eric M. Thompson <eric.thompson@tufts.edu>}
\examples{

}
\keyword{}
